StyleSwinGAN/requirements.txt
gdown
pytorch
torchvision
numpy
tqdm
scipy
matplotlib
timm
wandb
scikit-learn